direnv via externalfiles
When direnv is installed via homebrew, it automatically hooks in
fish (via linking activation script in /opt/homebrew/share/fish),
but in order to make it work nicely with rtx, order of hooks activation
is important. So, installing it manually and activating it manually
in config.fish with this change.

relevant ticket: https://github.com/jdxcode/rtx/issues/8